large amount of text changes
the bottom sections for the penis shrinking - those scenes still need to be hashed out. I wrote those a long time ago and they still aren't being used. Can a new scene be created for a while after the second penis shrinking talk with the therapist? I can write it, just need to know WHERE to write it.